progress on travis only

+++ build.sh
@@ -9,13 +9,21 @@ for a in "$@"; do
     fi
     repo=${d%/$tag}
     t="rosindustrial/${repo//\//-}:$tag" # replace other slashed with dashes
+
     echo
     echo "Build image $t"
-    while sleep 9m; do echo "Still building $t..."; done &
-    b=$!
+
+    if [ "$TRAVIS" ]; then
+      while sleep 9m; do echo "Still building $t..."; done &
+      b=$!
+    fi
+
     time docker build -q -t "$t" "$d" || ret=$?
-    disown
-    kill $b
+
+    if [ "$TRAVIS" ]; then
+      disown
+      kill $b
+    fi
   done
 done
 exit $ret
